+++ auction-service/src/main/java/com/dhernandez/auction_service/application/command/createAuctionCommand.java
@@ -1,5 +1,6 @@
 package com.dhernandez.auction_service.application.command;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class CreateAuctionCommand {
@@ -7,8 +8,8 @@ public class CreateAuctionCommand {
     private String description;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
-    private Double startingPrice;
-    public CreateAuctionCommand(String title, String description,  LocalDateTime startTime, LocalDateTime endTime, Double startingPrice){
+    private BigDecimal startingPrice;
+    public CreateAuctionCommand(String title, String description,  LocalDateTime startTime, LocalDateTime endTime, BigDecimal startingPrice){
         this.title = title;
         this.description = description;
         this.startTime = startTime;
@@ -27,7 +28,7 @@ public class CreateAuctionCommand {
     public LocalDateTime getEndTime(){
         return endTime;
     }
-    public Double getStartingPrice(){
+    public BigDecimal getStartingPrice(){
         return startingPrice;
     }   
 }
